Refactor LoginForm component
