+++ friends/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Route } from 'react-router-dom';
+import { Route, Link } from 'react-router-dom';
 import './App.css';
 import LoginForm from './components/LoginForm';
 import FriendList from './components/FriendList';
@@ -8,8 +8,16 @@ import PrivateRoute from './components/PrivateRoute';
 function App() {
   return (
     <div className="App">
+      <ul>
+        <li>
+          <Link to="/login">Login</Link>
+        </li>
+        <li>
+          <Link to="/protected">Friends List</Link>
+        </li>
+      </ul>
       <Route path="/login" component={LoginForm} />
-      <PrivateRoute path="/protected" component={FriendList} />
+      <PrivateRoute exact path="/protected" component={FriendList} />
     </div>
   );
 }
