+++ friends/src/components/FriendList.jsx
@@ -9,7 +9,7 @@ class FriendList extends Component {
 
   handleLogOut = () => {
     localStorage.removeItem('authToken');
-    this.props.history.push('/login');
+    this.props.history.push('/');
   }
 
   render() {
